Image reload logic.
If an image cannot be loaded, the flipbook will attempt to reload the images
for a configurable period of time (the default is 60 seconds). After this
time, any images that could not be loaded are marked as "not loaded".

If all the images are successfully loaded, a "flipbookload" event will be
triggered. If the images could not all be loaded, a "flipbookerror" event will
be triggered. Use the standard jQuery "bind" method to register your own
handlers for these events.